Move retry errors to shared dir
