+++ pkg/shared/ruleset/disak8sstig/retryerrors/retryerrors.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package retry
+package retryerrors
 
 import (
 	"regexp"
@@ -13,6 +13,6 @@ var (
 	ContainerNotFoundOnNodeRegexp = regexp.MustCompile(`(?i)(/var/lib/kubelet/pods.*No such file or directory)`)
 	// ContainerNotReadyRegexp regex to match container not yet in status or not running
 	ContainerNotReadyRegexp = regexp.MustCompile(`(?i)(container with name .* (not \(yet\) in status|not \(yet\) running))`)
-	// DikiDISAPodNotFoundRegexp regex to match diki pod not found for disa ruleset
-	DikiDISAPodNotFoundRegexp = regexp.MustCompile(`(?i)(pods "diki-[\d]{6}-.{10}" not found)`)
+	// OpsPodNotFoundRegexp regex to match ops pod not found for disa ruleset
+	OpsPodNotFoundRegexp = regexp.MustCompile(`(?i)(pods "diki-[\d]{6}-.{10}" not found)`)
 )
