+++ pkg/shared/ruleset/disak8sstig/retryerrors/retryerrors_suite_test.go
@@ -1,0 +1,17 @@
+// SPDX-FileCopyrightText: 2024 SAP SE or an SAP affiliate company and Gardener contributors
+//
+// SPDX-License-Identifier: Apache-2.0
+
+package retryerrors_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+func TestRetryErrors(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Retry Errors Test Suite")
+}
